requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.3.0
bottlenose==0.6.2
dateutils==0.6.6
itsdangerous==0.24
lxml==3.4.4
python-amazon-simple-product-api==2.0.1
python-dateutil==2.4.2
pytz==2015.4
six==1.9.0
wsgiref==0.1.2
